Add more assignment detail to README
I was using the test suite as a way to direct students through the
assignment, but I think this is more "cute" than productive. The
tests are hard to read for the students, and there is a lot for them
to get tripped on.

This commit adds the information that was in the test suite and adds
it to the "You will know you are done when..." section of the README.